% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbgc.R
\name{pbgc_context}
\alias{pbgc_context}
\title{Return context: full source of this client}
\usage{
pbgc_context()
}
\value{
character: source code of this file
}
\description{
Return context: full source of this client
}
