% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbgc.R
\name{pbgc_multiemployer_plans}
\alias{pbgc_multiemployer_plans}
\title{Fetch PBGC multiemployer pension plans}
\usage{
pbgc_multiemployer_plans(state = NULL)
}
\arguments{
\item{state}{Optional state abbreviation to filter}
}
\value{
tibble: plan_name, admin_city, admin_state, sponsor_name, phone,
effective_date, plan_id, ein, pn, participant_count
}
\description{
Fetch PBGC multiemployer pension plans
}
